+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 ## 0.23.0 _(2024-01-25)_
 
 -   You are able to send markdown code block in prompts _(without traces of escaping)_
+-   Function `trimEndOfCodeBlock` is not working with escaped code blocks JUST with markdown code blocks